# server/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Market Info API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/products/{product_id}": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Get Product",
        "operationId": "get_product_api_v1_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/merchant/dashboard": {
      "get": {
        "tags": [
          "merchant"
        ],
        "summary": "Get Dashboard",
        "operationId": "get_dashboard_api_v1_merchant_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/merchant/products": {
      "post": {
        "tags": [
          "merchant"
        ],
        "summary": "Create Product",
        "operationId": "create_product_api_v1_merchant_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/merchant/products/ai-draft": {
      "post": {
        "tags": [
          "merchant"
        ],
        "summary": "Ai Draft",
        "operationId": "ai_draft_api_v1_merchant_products_ai_draft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/merchant/products/{product_id}": {
      "patch": {
        "tags": [
          "merchant"
        ],
        "summary": "Update Product",
        "operationId": "update_product_api_v1_merchant_products__product_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/merchant/my-store": {
      "get": {
        "tags": [
          "merchant"
        ],
        "summary": "Get My Store",
        "operationId": "get_my_store_api_v1_merchant_my_store_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/merchant/drops/{drop_id}/status": {
      "patch": {
        "tags": [
          "merchant"
        ],
        "summary": "Update Drop Status",
        "operationId": "update_drop_status_api_v1_merchant_drops__drop_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "drop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Drop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DropStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/preorders": {
      "post": {
        "tags": [
          "preorders"
        ],
        "summary": "Create Preorder",
        "description": "사전 주문 생성 — 소비자·상인 모두 가능.",
        "operationId": "create_preorder_api_v1_preorders_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreorderCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "preorders"
        ],
        "summary": "List Preorders",
        "description": "내 사전 주문 목록.\n- consumer: 본인 주문만 조회\n- merchant: 담당 점포 주문 전체 조회\n- ?status= 로 상태 필터링 가능",
        "operationId": "list_preorders_api_v1_preorders_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "상태 필터: requested|confirmed|ready|cancelled",
              "title": "Status"
            },
            "description": "상태 필터: requested|confirmed|ready|cancelled"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/preorders/{preorder_id}": {
      "get": {
        "tags": [
          "preorders"
        ],
        "summary": "Get Preorder",
        "description": "단건 조회.\n- consumer: 본인 주문만 조회 가능\n- merchant: 담당 점포 주문만 조회 가능\n- 권한 외 접근 시 403",
        "operationId": "get_preorder_api_v1_preorders__preorder_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "preorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preorder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "preorders"
        ],
        "summary": "Cancel Preorder",
        "description": "소비자 직접 취소.\n- consumer 전용\n- requested 상태일 때만 가능 (그 외 → 409)\n- 취소 시 Notification 자동 생성",
        "operationId": "cancel_preorder_api_v1_preorders__preorder_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "preorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preorder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/merchant/preorders/{preorder_id}/status": {
      "patch": {
        "tags": [
          "preorders"
        ],
        "summary": "Update Preorder Status",
        "description": "Preorder 상태 변경 (merchant 전용).\n상태 변경 시 주문자에게 Notification 자동 생성.\n\n허용 전이:\n  requested → confirmed | cancelled\n  confirmed → ready     | cancelled\n  ready     → cancelled",
        "operationId": "update_preorder_status_api_v1_merchant_preorders__preorder_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "preorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preorder Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreorderStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIDraftRequest": {
        "properties": {
          "store_id": {
            "type": "string",
            "title": "Store Id"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "voice_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Text"
          }
        },
        "type": "object",
        "required": [
          "store_id"
        ],
        "title": "AIDraftRequest"
      },
      "BaseResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "요청이 성공했습니다."
          },
          "data": {
            "title": "Data"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Meta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "BaseResponse"
      },
      "DropStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "DropStatusUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Meta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "timestamp"
        ],
        "title": "Meta"
      },
      "PreorderCreate": {
        "properties": {
          "store_id": {
            "type": "string",
            "title": "Store Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "qty": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Qty"
          }
        },
        "type": "object",
        "required": [
          "store_id",
          "product_name",
          "qty"
        ],
        "title": "PreorderCreate"
      },
      "PreorderStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "PreorderStatusUpdate"
      },
      "ProductCreate": {
        "properties": {
          "store_id": {
            "type": "string",
            "title": "Store Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "price": {
            "type": "integer",
            "title": "Price"
          },
          "stock_status": {
            "type": "string",
            "title": "Stock Status"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "quality_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Note"
          }
        },
        "type": "object",
        "required": [
          "store_id",
          "product_name",
          "price",
          "stock_status"
        ],
        "title": "ProductCreate"
      },
      "ProductUpdate": {
        "properties": {
          "price": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "stock_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Status"
          }
        },
        "type": "object",
        "title": "ProductUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}